Resolves issues with printing hexadecimal with unsigned int limit

+++ main.h
@@ -27,7 +27,7 @@ int fill_buffer(char *buffer, long int num, int n, int base);
 int print_rev_buffer(char *buffer, int n);
 
 int _malloc3(va_list num_list, int base, char caps);
-int fill_buffer_hexa(char *buffer, int num, int n, int base, char caps);
+int fill_buffer_hexa(char *buffer, long int num, int n, int base, char caps);
 
 int rev_string(va_list num_list);
 int rot13(va_list num_list);
